requirements.txt
invoke
minio
pre-commit
PyYAML
python-dotenv
